Move log replay tests to separate file
